[Enhancement] Add compaction memory statistics in memtracker for cloud native tablet (#26843)
