+++ be/src/storage/CMakeLists.txt
@@ -170,6 +170,7 @@ add_library(Storage STATIC
     lake/async_delta_writer.cpp
     lake/compaction_policy.cpp
     lake/compaction_scheduler.cpp
+    lake/compaction_task.cpp
     lake/horizontal_compaction_task.cpp
     lake/delta_writer.cpp
     lake/vacuum.cpp
