+++ be/src/storage/lake/compaction_task.cpp
@@ -12,33 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
-
-#include <memory>
-#include <vector>
-
 #include "storage/lake/compaction_task.h"
 
-namespace starrocks {
-class Chunk;
-class ChunkIterator;
-} // namespace starrocks
+#include "storage/lake/tablet.h"
 
 namespace starrocks::lake {
 
-class TabletWriter;
-
-class HorizontalCompactionTask : public CompactionTask {
-public:
-    explicit HorizontalCompactionTask(int64_t txn_id, int64_t version, std::shared_ptr<Tablet> tablet,
-                                      std::vector<std::shared_ptr<Rowset>> input_rowsets)
-            : CompactionTask(txn_id, version, std::move(tablet), std::move(input_rowsets)) {}
-    ~HorizontalCompactionTask() override = default;
-
-    Status execute(Progress* progress, CancelFunc cancel_func) override;
-
-private:
-    StatusOr<int32_t> calculate_chunk_size();
-};
+CompactionTask::CompactionTask(int64_t txn_id, int64_t version, std::shared_ptr<Tablet> tablet,
+                               std::vector<std::shared_ptr<Rowset>> input_rowsets)
+        : _txn_id(txn_id),
+          _version(version),
+          _tablet(std::move(tablet)),
+          _input_rowsets(std::move(input_rowsets)),
+          _mem_tracker(std::make_unique<MemTracker>(MemTracker::COMPACTION, -1,
+                                                    "Compaction-" + std::to_string(_tablet->id()),
+                                                    ExecEnv::GetInstance()->compaction_mem_tracker())) {}
 
 } // namespace starrocks::lake
