+++ be/src/storage/lake/compaction_task.h
@@ -19,8 +19,12 @@
 #include <ostream>
 
 #include "common/status.h"
+#include "runtime/mem_tracker.h"
 
 namespace starrocks::lake {
+
+class Rowset;
+class Tablet;
 
 class CompactionTask {
 public:
@@ -38,12 +42,21 @@ public:
     // should be cancelled.
     using CancelFunc = std::function<bool()>;
 
+    explicit CompactionTask(int64_t txn_id, int64_t version, std::shared_ptr<Tablet> tablet,
+                            std::vector<std::shared_ptr<Rowset>> input_rowsets);
     virtual ~CompactionTask() = default;
 
     virtual Status execute(Progress* stats, CancelFunc cancel_func) = 0;
 
     inline static const CancelFunc kNoCancelFn = []() { return false; };
     inline static const CancelFunc kCancelledFn = []() { return true; };
+
+protected:
+    int64_t _txn_id;
+    int64_t _version;
+    std::shared_ptr<Tablet> _tablet = nullptr;
+    std::vector<std::shared_ptr<Rowset>> _input_rowsets;
+    std::unique_ptr<MemTracker> _mem_tracker = nullptr;
 };
 
 } // namespace starrocks::lake
