+++ be/src/storage/lake/horizontal_compaction_task.cpp
@@ -32,6 +32,9 @@ Status HorizontalCompactionTask::execute(Progress* progress, CancelFunc cancel_f
     if (progress == nullptr) {
         return Status::InvalidArgument("progress is null");
     }
+
+    SCOPED_THREAD_LOCAL_MEM_TRACKER_SETTER(_mem_tracker.get());
+
     ASSIGN_OR_RETURN(auto tablet_schema, _tablet->get_schema());
     int64_t total_num_rows = 0;
     for (auto& rowset : _input_rowsets) {
