+++ be/src/storage/lake/vertical_compaction_task.h
@@ -29,19 +29,13 @@ struct RowSourceMask;
 
 namespace starrocks::lake {
 
-class Rowset;
-class Tablet;
 class TabletWriter;
 
 class VerticalCompactionTask : public CompactionTask {
 public:
     explicit VerticalCompactionTask(int64_t txn_id, int64_t version, std::shared_ptr<Tablet> tablet,
                                     std::vector<std::shared_ptr<Rowset>> input_rowsets)
-            : _txn_id(txn_id),
-              _version(version),
-              _tablet(std::move(tablet)),
-              _input_rowsets(std::move(input_rowsets)) {}
-
+            : CompactionTask(txn_id, version, std::move(tablet), std::move(input_rowsets)) {}
     ~VerticalCompactionTask() override = default;
 
     Status execute(Progress* progress, CancelFunc cancel_func) override;
@@ -54,11 +48,7 @@ private:
                                 RowSourceMaskBuffer* mask_buffer, std::vector<RowSourceMask>* source_masks,
                                 Progress* progress, const CancelFunc& cancel_func);
 
-    int64_t _txn_id;
-    int64_t _version;
-    std::shared_ptr<Tablet> _tablet;
     std::shared_ptr<const TabletSchema> _tablet_schema;
-    std::vector<std::shared_ptr<Rowset>> _input_rowsets;
     int64_t _total_num_rows = 0;
     int64_t _total_data_size = 0;
     int64_t _total_input_segs = 0;
